fix: update about us page content
