+++ src/config/partners.ts
@@ -1,0 +1,44 @@
+export const partners = [
+	{ name: "Adobe", imageSrc: "/Adobe.webp" },
+	{ name: "Goldman-sachs", imageSrc: "/Goldman-Sachs.webp" },
+	{
+		name: "Government-of-Madhya-Pradesh",
+		imageSrc: "/Government-of-Madhya-Pradesh.webp",
+	},
+	{ name: "MalalaFund", imageSrc: "/MalalaFund.webp" },
+	{ name: "NSIF-Nasscom-Awards", imageSrc: "/NSIF-Nasscom-Awards.webp" },
+	{ name: "Skoll-Foundation", imageSrc: "/Skoll-Foundation.webp" },
+	{ name: "Stir-Education", imageSrc: "/Stir-Education.webp" },
+	{
+		name: "UBS-Opitmus-Foundation",
+		imageSrc: "/UBS-Opitmus-Foundation.webp",
+	},
+	{
+		name: "Ray-and-Tye-Noorda-Foundation",
+		imageSrc: "/Ray-and-Tye-Noorda-Foundation.webp",
+	},
+	{
+		name: "MIT_Solves_Learning_for_Girls_Women_Challenge_2020",
+		imageSrc: "/MIT_Solves_Learning_for_Girls_Women_Challenge_2020.webp",
+	},
+	{ name: "milton", imageSrc: "/milton.webp" },
+	{ name: "Marico", imageSrc: "/Marico.webp" },
+	{
+		name: "LOreal-Paris-Women-of-Worth",
+		imageSrc: "/LOreal-Paris-Women-of-Worth.webp",
+	},
+	{ name: "LGT", imageSrc: "/LGT.webp" },
+	{ name: "Jester-Foundation", imageSrc: "/ester-Foundation.webp" },
+	{ name: "jasmine-1", imageSrc: "/jasmine-1.webp" },
+	{ name: "ITC-BLUE", imageSrc: "/ITC-BLUE.webp" },
+	{ name: "iPartner-Indian", imageSrc: "/iPartner-India.webp" },
+	{ name: "icon-330", imageSrc: "/icon-330.webp" },
+	{
+		name: "HundrED_Sticker-pink-1",
+		imageSrc: "/HundrED_Sticker-pink-1.webp",
+	},
+	{
+		name: "Government-of-Uttar-Pradesh",
+		imageSrc: "/Government-of-Uttar-Pradesh.webp",
+	},
+];
