+++ src/config/team.ts
@@ -1,0 +1,26 @@
+export const teamembers = [
+	{
+		name: "Mr. Shivam Pathak",
+		role: "Director",
+		imageSrc: "/man.png",
+		linkedinUrl: "https://www.linkedin.com/in/ShivamPathak",
+	},
+	{
+		name: "Mrs. Pooja Pathakk",
+		role: "Director",
+		imageSrc: "/woman.png",
+		linkedinUrl: "https://www.linkedin.com/in/PoojaPathak",
+	},
+	{
+		name: "Mr. Nilesh Pal",
+		role: "Director",
+		imageSrc: "/man.png",
+		linkedinUrl: "https://www.linkedin.com/in/NileshPathak",
+	},
+	{
+		name: "Mr. Zeel Mangukiya",
+		role: "COO",
+		imageSrc: "/man.png",
+		linkedinUrl: "https://www.linkedin.com/in/ZeelMangukiya",
+	},
+];
